+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/ast/AllJavaAstTreeDumpTest.java
@@ -14,8 +14,8 @@ import org.junit.platform.suite.api.Suite;
     Java15TreeDumpTest.class,
     Java16TreeDumpTest.class,
     Java17TreeDumpTest.class,
-    Java18PreviewTreeDumpTest.class,
-    Java19PreviewTreeDumpTest.class
+    Java19PreviewTreeDumpTest.class,
+    Java20PreviewTreeDumpTest.class
 })
 class AllJavaAstTreeDumpTest {