+++ pmd-java/src/test/kotlin/net/sourceforge/pmd/lang/java/ast/KotlinTestingDsl.kt
@@ -31,8 +31,9 @@ enum class JavaVersion : Comparable<JavaVersion> {
     J15,
     J16,
     J17,
-    J18, J18__PREVIEW,
-    J19, J19__PREVIEW;
+    J18,
+    J19, J19__PREVIEW,
+    J20, J20__PREVIEW;
 
     /** Name suitable for use with e.g. [JavaParsingHelper.parse] */
     val pmdName: String = name.removePrefix("J").replaceFirst("__", "-").replace('_', '.').lowercase()